backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Git operations
gitpython==3.1.41

# Vector store
chromadb==0.4.22

# OpenAI
openai==1.12.0
google-generativeai==0.8.4

# Language detection
pygments==2.17.2

# HTTP client
httpx==0.26.0

# Utilities
tiktoken==0.5.2
tenacity==8.2.3
structlog==24.1.0
